requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10
gunicorn==23.0.0
requests==2.32.3
python-dotenv==1.0.1
qrcode[pil]==7.4.2
